Added POM structure to the project
