+++ AutomationExerciseSuite/src/test/java/pageObject/LoginPage.java
@@ -1,0 +1,27 @@
+package pageObject;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class LoginPage {
+	WebDriver driver;
+	public LoginPage(WebDriver driver) {
+		this.driver = driver;
+	}
+	By txt_username=By.xpath("(//input[@type='email' and @name='email'])[1]");
+	By txt_password=By.xpath("//input[@placeholder='Password']");
+	By btn_login=By.xpath("//button[contains(text(), 'Login')]");
+	
+	public void enterUsername(String username) {
+		driver.findElement(txt_username).sendKeys(username);
+	}
+	public void enterPassword(String password) {
+		driver.findElement(txt_password).sendKeys(password);
+	}
+	
+	public void clickLoginButton() {
+		driver.findElement(btn_login).click();
+	}
+	
+
+}
